test: introduce helper files to separate helper functions from actual tests
